glasm: Move implemented instructions from not_implemented.cpp
